Refactor retry logic
* Less procs & proc building overall
* seperate delay calculation into its own class
