+++ spec/lib/http/retriable/performer_spec.rb
@@ -5,6 +5,7 @@ RSpec.describe HTTP::Retriable::Performer do
   let(:client) do
     HTTP::Client.new
   end
+
   let(:response) do
     HTTP::Response.new(
       :status  => 200,
@@ -15,6 +16,7 @@ RSpec.describe HTTP::Retriable::Performer do
       :request => request
     )
   end
+
   let(:request) do
     HTTP::Request.new(
       :verb => :get,
@@ -73,8 +75,7 @@ RSpec.describe HTTP::Retriable::Performer do
     end
 
     describe "expected status codes" do
-
-      def response(options = {})
+      def response(**options)
         HTTP::Response.new(
           {
             :status  => 200,
@@ -96,14 +97,13 @@ RSpec.describe HTTP::Retriable::Performer do
       end
 
       describe "status codes can be expressed in many ways" do
-
         [
           301,
           [200, 301, 485],
           250...400,
           [250...Float::INFINITY],
-          ->(status_code) { status_code == 301  },
-          [->(status_code) { status_code == 301  }]
+          ->(status_code) { status_code == 301 },
+          [->(status_code) { status_code == 301 }]
         ].each do |retry_statuses|
           it retry_statuses.to_s do
             expect do
@@ -302,63 +302,6 @@ RSpec.describe HTTP::Retriable::Performer do
         err = e
       end
       expect(err.response).to be response
-    end
-  end
-
-  describe "delay calculation" do
-    def call_delay(iterations, **options)
-      HTTP::Retriable::Performer.new(options).calculate_delay_from_iteration(iterations)
-    end
-
-    def call_retry_header(value, **options)
-      HTTP::Retriable::Performer.new(options).delay_from_retry_header(value)
-    end
-
-    it "prevents negative sleep time" do
-      expect(call_delay(20, :delay => -20)).to eq 0
-    end
-
-    it "backs off exponentially" do
-      expect(call_delay(1)).to be_between 0, 1
-      expect(call_delay(2)).to be_between 1, 2
-      expect(call_delay(3)).to be_between 3, 4
-      expect(call_delay(4)).to be_between 7, 8
-      expect(call_delay(5)).to be_between 15, 16
-    end
-
-    it "can have a maximum wait time" do
-      expect(call_delay(1, :max_delay => 5)).to be_between 0, 1
-      expect(call_delay(5, :max_delay => 5)).to eq 5
-    end
-
-    it "respects Retry-After headers as integer" do
-      delay_time = rand(6...2500)
-      header_value = delay_time.to_s
-      expect(call_retry_header(header_value)).to eq delay_time
-      expect(call_retry_header(header_value, :max_delay => 5)).to eq 5
-    end
-
-    it "respects Retry-After headers as rfc2822 timestamp" do
-      delay_time = rand(6...2500)
-      header_value = (Time.now.gmtime + delay_time).to_datetime.rfc2822.sub("+0000", "GMT")
-      expect(call_retry_header(header_value)).to be_within(1).of(delay_time)
-      expect(call_retry_header(header_value, :max_delay => 5)).to eq 5
-    end
-
-    it "respects Retry-After headers as rfc2822 timestamp in the past" do
-      delay_time = rand(6...2500)
-      header_value = (Time.now.gmtime - delay_time).to_datetime.rfc2822.sub("+0000", "GMT")
-      expect(call_retry_header(header_value)).to eq 0
-    end
-
-    it "does not error on invalid Retry-After header" do
-      [ # invalid strings
-        "This is a string with a number 5 in it",
-        "8 Eight is the first digit in this string",
-        "This is a string with a #{Time.now.gmtime.to_datetime.rfc2822} timestamp in it"
-      ].each do |header_value|
-        expect(call_retry_header(header_value)).to eq 0
-      end
     end
   end
 end
